+++ Adj/Program/Controlflow/Implementation/State.hs
@@ -1,12 +1,12 @@
 module Adj.Program.Controlflow.Implementation.State where
 
-import Adj.Algebra.Category ((.:), (:*:) ((:*:)))
+import Adj.Algebra.Category ((.), (.:), type (-->), Flat (Flat), (:*:) ((:*:)), type (:*:>))
 import Adj.Auxiliary (type (=!?=), FG (FG))
 
-type State state = (->) state =!?= (:*:) state
+type State state = (-->) state =!?= (:*:>) state
 
 current :: State state state
-current = FG .: \state -> state :*: state
+current = FG . Flat .: \state -> Flat (state :*: state)
 
 modify :: (state -> state) -> State state state
-modify m = FG .: \state -> m state :*: state
+modify m = FG . Flat .: \state -> Flat (m state :*: state)