+++ Adj/Program/Controlflow/Implementation/Store.hs
@@ -1,12 +1,12 @@
 module Adj.Program.Controlflow.Implementation.Store where
 
-import Adj.Algebra.Category ((:*:) ((:*:)), (...:))
+import Adj.Algebra.Category (type (-->), type (:*:>), (.), (:*:) ((:*:)), Flat (Flat), (...:))
 import Adj.Auxiliary (type (=!?=), FG (FG))
 
-type Store store = (:*:) store =!?= (->) store
+type Store store = (:*:>) store =!?= (-->) store
 
 position :: Store store o -> store
-position (FG (store :*: _)) = store
+position (FG (Flat (store :*: _))) = store
 
 retrofit :: (store -> store) -> Store store store -> Store store store
-retrofit g (FG (store :*: f)) = FG ...: g store :*: f
+retrofit g (FG (Flat (store :*: f))) = FG . Flat ...: g store :*: f